Allocate instances via kv instead of services
Previously consul services were used to set the allocation of
each individual instance. But it turns out that services are
not unique across consul nodes. So in case of node failure,
the system may end up with 2 service copies when the node goes
up again.

Thus the strongly-consistent kv storage is chosen instead.